==> src/notebooks-python/archive_team_analysis/02_archive_team_hashtag_sampling.ipynb <==
import pickle
import numpy as np
import pandas as pd
from functools import reduce
import os

Get list of paths to files

listOfFiles = list()
for (dirpath, dirnames, filenames) in os.walk(r"G:\TwitterUsers\saved_objs_withKeywords"):
    listOfFiles += [os.path.join(dirpath, file) for file in filenames]

#listOfFiles = listOfFiles[-47:-16]

len(listOfFiles)

Set initial hashtags to look for

initial_keywords = set(["election2012","vote","voteblue","votered","obama","romney"])

Perform first search

%%time
hashtags = set()
i = 0
total = len(listOfFiles)
save_path = r"G:\TwitterUsers\processed_files\hashtags\hashtags_1.txt" 
backup_save_path = r"G:\TwitterUsers\processed_files\hashtags\backup\hashtags_1.txt" 
for path in listOfFiles:
    print("        doing", i, "-th file out of ", total, end = "\r")
    with open(path,"rb") as f:
        day = pickle.load(f)
        f.close()
#     j = 0
#     tot_tweets = len(day)
    print("loaded", end = "\r")
    for tweet in day:
        #if not set([hashtag.lower() for hashtag in tweet["hashtags"]]).isdisjoint(initial_hashtag):
        if any(kw in tweet["text"] for kw in initial_keywords ):
            hashtags.update(tweet["hashtags"])
        
#         j += 1
#         if j%10000 == 0:
#             print("done,",j," tweets out of", tot_tweets, "of",i,"-th file out of", total, end = "\r")
    
    i += 1
    print("saving", end = "\r")
    with open(save_path, "wb") as f:
        pickle.dump(list(hashtags),f)
        f.close()
        
    with open(backup_save_path, "wb") as f:
        pickle.dump(list(hashtags),f)
        f.close()

Load hashtags resulting from first search

save_path = r"G:\TwitterUsers\processed_files\hashtags\hashtags_1.txt" 
with open(save_path, "rb") as f:
    hashtags = pickle.load(f)
    f.close()
print(len(hashtags))

Set interesting pieces to look for in the hastags resulting from first search.

pieces = set(["democrat","republican","obama","romney","votered","voteblue"])

Select second round hashtags

second_hashtags = set([hashtag for hashtag in hashtags  for piece in pieces if piece in hashtag.lower()])
print(len(second_hashtags)) #1010
#list(second_hashtags)[:100]

list(second_hashtags)[:100]

Perform second hashtag search

%%time
hashtags = set()
i = 0
total = len(listOfFiles)
save_path = r"G:\TwitterUsers\processed_files\hashtags\hashtags_2.txt" 
backup_save_path = r"G:\TwitterUsers\processed_files\hashtags\backup\hashtags_2.txt" 
for path in listOfFiles:
    print("        doing", i, "-th file out of ", total, end = "\r")
    with open(path,"rb") as f:
        day = pickle.load(f)
        f.close()
#     j = 0
#     tot_tweets = len(day)
    print("loaded", end = "\r")
    for tweet in day:
        #if not set([hashtag.lower() for hashtag in tweet["hashtags"]]).isdisjoint(second_hashtags):
        if any(kw in tweet["text"] for kw in second_hashtags ):
            hashtags.update(tweet["hashtags"])
        
#         j += 1
#         if j%10000 == 0:
#             print("done,",j," tweets out of", tot_tweets, "of",i,"-th file out of", total, end = "\r")
    
    i += 1
    print("saving", end = "\r")
    with open(save_path, "wb") as f:
        pickle.dump(list(hashtags),f)
        f.close()
        
    with open(backup_save_path, "wb") as f:
        pickle.dump(list(hashtags),f)
        f.close()

Load second round hashtags

with open(r"G:\TwitterUsers\processed_files\hashtags\hashtags_2.txt", "rb") as f:
    hashtags = pickle.load(f)
    f.close()
    
print(len(hashtags))

Extract interesting hashtags and save file

third_hashtags =  list(set([hashtag for hashtag in hashtags  for piece in pieces if piece in hashtag.lower()]))
print(len(third_hashtags))

with open(r"G:\TwitterUsers\processed_files\hashtags\hashtags_3.txt","wb") as f:
    pickle.dump(third_hashtags,f)
    f.close()
