==> src/notebooks-python/archive_team_analysis/04_sampling_from_archive_team.ipynb <==
import sys
sys.path.insert(1, r"Q:\IlMIoDrive\magistrale\1anno\terzo_periodo\MAS\analysis")
#sys.path.insert(1, r"C:\Econophysics\analysis")
from importlib import reload
import Econophysics as Ec
import credentials as cr
Ec = reload(Ec)
cr  = reload(cr)
import pickle
from threading import Thread
from datetime import datetime
import tweepy
import os
import numpy as np
import pandas as pd

len(cr.apps) #https://archive.org/details/twitterstream?&sort=-publicdate

Load retweet network extracted from hashtag mining

%%time
#retweet_network_np = np.load(r"F:\TwitterUsers\archiveTeam\retweet_network\retweet_network1.npy", allow_pickle = True)
retweet_network_np = np.load(r"G:\TwitterUsers\processed_files\retweet_network\retweet_network.npy", allow_pickle = True)
retweet_network_csv = pd.DataFrame(retweet_network_np,columns = ["user_id",	"profile_created_at", "retweeted_tweet_id", "retweeted_user_id", "hashtags"] )
retweet_network = retweet_network_csv.to_dict("records")
retweet_network[:3]

Check that users are uniques (which means validate previous parts of the pipeline)

len(set([user["user_id"] for user in retweet_network])) == len(retweet_network)

%%time
#Ec = reload(Ec) #started at 00:40
#path_to_dir = r"F:\TwitterUsers\archiveTeam\saved_objs\05\\"
threads = Ec.get_AT_threads(n_threads = 20,users = retweet_network ,apps = cr.apps, resource =["resources","users","/users/:id"] , save_path = r"G:\TwitterUsers\processed_files\AT_thread_output" , restart = False)
#thread = AT_Thread(name = "thread0",paths =[path_to_dir+"objs_1440.txt",path_to_dir+"objs_2880.txt"] ,apps=cr.apps,resource_remaining=["resources","users","/users/:id","remaining"],resource_reset=["resources","users","/users/:id","reset"],save_path = r"F:\TwitterUsers\archiveTeam\processed_files",restart=False)

%%time
Ec.start_threads(threads, wait = 0) #started at 19:29
