==> src/notebooks-python/archive_team_analysis/05_initialization_RTWT.ipynb <==
import sys
sys.path.insert(1, r"Q:\IlMIoDrive\magistrale\1anno\terzo_periodo\MAS\analysis")
from importlib import reload
import credentials as cr
import Econophysics as Ec
Ec = reload(Ec)
cr = reload(cr)
import tweepy
from threading import Thread
import numpy as np
import os
import pickle
from datetime import datetime,timezone
import time

from collections import Counter
import networkx as nx
import matplotlib.pyplot as plt 
from strawberryfields.apps import subgraph

with open(r"G:\TwitterUsers\processed_files\AT_thread_output\ok_users.txt", "rb") as f: #C:\Econophysics\TwitterUsers\archiveTeam\processed_files  # r"F:\TwitterUsers\archiveTeam\processed_files\ok_users.txt"
    ok_users = pickle.load(f)
    f.close()
    
print(len(ok_users))

name = "Initialization"
path_start_rtwt = r"G:\TwitterUsers\processed_files\Initialization_RTWT_output_mentions_up\threading_objs\\" +name+"_retweet_objs_" 
rtwt_threads = Ec.get_threads(n_threads = 30, thread = Ec.TwitterThreadRTWT, path_start = path_start_rtwt , path_end = ".txt", ids  = ok_users, apps = cr.apps,resource = ["resources","statuses",'/statuses/user_timeline'],restart  = False)

%%time
Ec.start_threads(rtwt_threads)

Ec.objs_txt_unifier(threading_folder =r"G:\TwitterUsers\processed_files\Initialization_RTWT_output_mentions_up\threading_objs" , destination_folder = r"G:\TwitterUsers\processed_files\Initialization_RTWT_output_mentions_up", name = "Initialization_retweet_network_mentions_up_14_10",mode = "a" )

Check

objs = np.load(r"G:\TwitterUsers\processed_files\Initialization_RTWT_output_mentions_up\Initialization_retweet_network_mentions_up_14_10.npy", allow_pickle = True)
len(objs)
#objs[0]

## Basic Analysis

# retweet_network = np.load(r"F:\TwitterUsers\archiveTeam\Streamer_archiveTeam\user_timeline\Initialization_retweet_network.npy", allow_pickle = True)
# print(len(retweet_network))
# nodes = set([user["user_id"] for user in retweet_network])
# print(len(nodes))
# all_edges = [(user["user_id"], user["retweeted_users_ids"][i]) for user in retweet_network for i in range(len(user["retweeted_users_ids"])) ] 
# print(len(all_edges))
# print(all_edges[0])
# edge_list = [tup for tup in all_edges if tup[1] in nodes]
# len(edge_list)

# %%time
# retweet_counter = dict(Counter(edge_list))
# weighted_edge_list = [(*list(retweet_counter.keys())[i], list(retweet_counter.values())[i]) for i in range(len(retweet_counter ))]

# retweet_network = nx.DiGraph()
# #retweet_network.add_nodes_from(nodes)
# retweet_network.add_weighted_edges_from(weighted_edge_list)
# retweet_network.name = "retweet network"
# # Show the basic attributes of the graph
# print(nx.info(retweet_network))

# %%time
# dense1 = subgraph.search([retweet_network.to_undirected()], retweet_network.to_undirected(), 100, 101, max_count=30)
# dense1
# sub_100 = retweet_network.subgraph(dense1[100][0][1])
# with open(r"F:\TwitterUsers\archiveTeam\Streamer_archiveTeam\densest\densest.txt", "wb") as f:
#     pickle.dump(dense1,f)
#     f.close()
# type(dense1)
# dense2 = pickle.load(r"F:\TwitterUsers\archiveTeam\Streamer_archiveTeam\densest\densest.txt")
# my_dense =[]
# for boh in dense:
#     my_dense.append(boh)
# my_dense
