==> src/notebooks-python/archive_team_analysis/07_2_tweets_ids_binary_processing.ipynb <==
import sys
#sys.path.insert(1, r"/Users/Pit/GitHub/Econophysics/Project/TwitterMonitor/")
sys.path.insert(1, r"E:\IlMIoDrive\magistrale\1anno\terzo_periodo\MAS\analysis")
import pickle
import pandas as pd
import numpy as np
from importlib import reload
import Econophysics as Ec
import os
from collections import Counter
Ec=reload(Ec)

paths = Ec.get_listOfFiles(r"F:\TwitterUsers\saved_objs_withKeywords\01")

day = Ec.pickle_loader(paths[0])

bin(day[0]["tweet_id"])

"0123456789101112"[:-1]

day[0]

1325397600000 -1288834974657

 1325401200000 - 1288834974657

def get_t_mId_seq(tweet_id):
    binary = bin(tweet_id)
    return (eval(binary[:-22]),eval("0b"+binary[-22:-12]),eval("0b"+binary[-12:]))

def get_machineId_sequence(tweet_id):
    return (eval("0b"+bin(tweet_id)[-22:-12]),eval("0b"+bin(tweet_id)[-12:]))

def get_timestamp(tweet_id):
    return eval(bin(tweet_id)[:-22])

machine_seq  = []
for tweet in day:
    machine_seq = [get_machineId_sequence(tweet["tweet_id"]) for tweet in ]
    
timestamps = [get_timestamp(tweet["tweet_id"]) for tweet in day]

len(np.unique([couple[0] for couple in machine_seq]))

np.mean([couple[1] for couple in machine_seq])

np.std([couple[1] for couple in machine_seq])

%%time
mp = map(lambda x: (eval(bin(x)[:-22]),eval("0b"+bin(x)[-22:-12]),eval("0b"+bin(x)[-12:])) , day)#[ (eval(bin(tweet_id)[:-22]),eval("0b"+bin(tweet_id)[-22:-12]),eval("0b"+bin(tweet_id)[-12:])) for tweet_id in day]

%%time
tot = sum([1 for x in mp])

%%time
lst = list(mp)

str(lst[0][0])[-3:-1]

%%time

# get metadata
start_path = r"F:\TwitterUsers\final_analysis\create_hashtagged_tweets\2012_10000_grab\all_daily_ids\\"
dirs = [start_path + path for path in  os.listdir(start_path)]

for j,dirct in enumerate(dirs):
    
    paths   = Ec.get_listOfFiles(dirct)
    
    metadatum = dict()
    t_mId_seq = []
    total_days = len(paths)
    for i,path in enumerate(paths):
        day = Ec.pickle_loader(path)
        #t_mId_seq.append(map(lambda x: (eval(bin(x)[:-22]),eval("0b"+bin(x)[-22:-12]),eval("0b"+bin(x)[-12:])) , day))
        t_mId_seq.extend([ (eval(bin(tweet_id)[:-22]),eval("0b"+bin(tweet_id)[-22:-12]),eval("0b"+bin(tweet_id)[-12:])) for tweet_id in day])
        print("done",i,"-th day out of",total_days, "within",j,"-th month out of",str(11),"months", end = "\r" )
    print("extracting timestamps...", end = "\r")
    timestamps = [triple[0] for triple in t_mId_seq]
    #timestamps = [triple[0] for mp in t_mId_seq for triple in mp]
    print("timestamps_delta_regularity...", end = "\r")
    metadatum["timestamps_delta_regularity"] = [max(timestamps) , min(timestamps), sum([1 if str(timestamp)[-3:-1] != "00" else 0 for timestamp in timestamps])]
    print("unique_machines...", end = "\r")
    metadatum["unique_machines"]  = dict(Counter([triple[1] for triple in t_mId_seq])) #list(set([triple[1] for triple in t_mId_seq]))
    print("sequences_counter...", end = "\r")
    metadatum["sequences_counter"] = dict(Counter([triple[2] for triple in t_mId_seq])) #[np.mean()]
    
    Ec.pickle_dumper(r"D:\TwitterUsers\archiveTeam\final_analysis\create_hashtagged_tweets\2012_10000_grab\metadata\metadatum_"+dirct[-2:]+".txt",metadatum)
    

paths = Ec.get_listOfFiles(r"F:\TwitterUsers\final_analysis\create_hashtagged_tweets\2012_10000_grab\all_daily_ids\01")
t_mId_seq = []
day = Ec.pickle_loader(paths[0])
#t_mId_seq.append(map(lambda x: (eval(bin(x)[:-22]),eval("0b"+bin(x)[-22:-12]),eval("0b"+bin(x)[-12:])) , day))
t_mId_seq.extend([ (eval(bin(tweet_id)[:-22]),eval("0b"+bin(tweet_id)[-22:-12]),eval("0b"+bin(tweet_id)[-12:])) for tweet_id in day])

sorted_t_mId_seq = sorted(t_mId_seq, key= lambda tup: tup[0])

t_seq_unique = np.unique([(tup[0],tup[2]) for tup in t_mId_seq])

len(t_mId_seq)

len(t_mId_seq) - len(t_seq_unique)

sorted_t_mId_seq[1:50]

metadata = []
for path in Ec.get_listOfFiles(r"D:\TwitterUsers\archiveTeam\final_analysis\create_hashtagged_tweets\2012_10000_grab\metadata"):
    metadata.append(Ec.pickle_loader(path))

#evaluate mean sequence value
mean_seq = np.mean([np.sum([key*value for key,value in metadatum["sequences_counter"].items()])/np.sum([value for value in metadatum["sequences_counter"].values() ]) for metadatum in metadata])
mean_seq

unique_machines = np.unique([machine for metadatum in metadata for machine in metadatum["unique_machines"].keys()])
unique_machines

unique_machines_all = [metadatum["unique_machines"] for metadatum in metadata]
unique_machines_all

[path for path in os.listdir(".")]

os.listdir(".")

bin(36)

machine_seq[:40]

timestamps[:40]
