==> src/notebooks-python/archive_team_analysis/08_hashtags_annotator.ipynb <==
# Hashtags Annotator

import pickle
from collections import OrderedDict
import ast
from IPython.display import clear_output
import warnings
import sys
sys.path.insert(1, r"E:\IlMIoDrive\magistrale\1anno\terzo_periodo\MAS\analysis")

import Econophysics as Ec

from importlib import reload
Ec = reload(Ec)

Here we construct the annotator object (implemented in Econophysics module) that launches the relative app. <br> <br>
Before running this cell, make sure you created a folder containing another folder named `backup`. As you see in the code below, I created the `my_folder` inside `D:\TwitterUsers\archiveTeam\final_analysis\hashtags` and inside `my_folder` I also created the `backup` folder. <br>
Please *copy* your txt file containing the list of dicts inside that folder (make a copy, don't cut and paste it for backup reasons). <br>
Once you did it, assign the path to your `my_folder` to the **folder** variable and the path to the [hashtagged_tweets.txt](https://drive.google.com/file/d/15rU3b_bi6tErOL_rU0v4-kBz-SWil93F/view?usp=sharing) to the **tweets_path** variable , then run the following cell . Plase follow the instructions presented as you proceed. 

Important: Should the app or jupyter crash, just do a backup of the hashtag_{name}.txt and the info.txt inside `my_folder`, restart jupyter and re-run al cells. It should work.

# insert path to my_folder here
folder = r"D:\TwitterUsers\archiveTeam\final_analysis\hashtags\Annotator_folder\\"
tweets_path = r"D:\TwitterUsers\archiveTeam\final_analysis\create_hashtagged_tweets\hashtagged_tweets_AT_38M_10000.txt" 
# get annotator object.
# who : insert your name, lower case
# save_path: where hashtags will be saved
# tweets_path: path to hashtagged_tweets file. You may find it here: https://drive.google.com/drive/u/7/folders/1Nfrl_P3GeQn4GAobfCWQ1ag75uMvjm_e
H = Ec.hashtag_annotator(who = "claudio",save_path = folder, backup_path =folder +r"backup\\" ,tweets_path = tweets_path)

# run the app
H.display_and_polarize()
